# Server/openapi.yaml
openapi: 3.0.0
info:
  title: TestVar - Flashcards API
  description: "A revolutionary REST API for flashcards"
  version: 1.0.0
servers:
  - url: http://localhost:7222/api
    description: Development server
paths:
...
  /sets:
    get:
      summary: "Get all flashcard sets"
      tags:
        - "FlashCard Sets"
      responses:
        "200":
          description: "A list of flashcard sets"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FlashcardSet"
    post:
      summary: "Create a new flashcard set"
      description: "There can only be 20 new flashcard sets created per day"
      tags:
        - "FlashCard Sets"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardSet"
      responses:
        "201":
          description: "The created flashcard set"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardSet"
        "429":
          description: "You have reached the maximum number of flashcard sets allowed today"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /sets/{setId}:
    parameters:
      - name: setId
        in: path
        required: true
        description: "The ID of the flashcard set"
        schema:
          type: "string"
    get:
      summary: "Get a flashcard set by ID"
      tags:
        - "FlashCard Sets"
      responses:
        "200":
          description: "The flashcard set"
          content:
            application/json:
              schema:
                type: "object"
                allOf:
                  - $ref: "#/components/schemas/FlashcardSet"
                  - type: "object"
                    properties:
                      comments:
                        type: "array"
                        description: "Comments on the set"
                        items:
                          type: "object"
                          properties:
                            comment:
                              type: "string"
                              example: "I love this set!"
                            author:
                              $ref: "#/components/schemas/User"
        "404":
          description: "The flashcard set was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    put:
      summary: "Update a flashcard set by ID"
      tags:
        - "FlashCard Sets"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardSet"
      responses:
        "200":
          description: "The updated flashcard set"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardSet"
        "404":
          description: "The flashcard set was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    delete:
      summary: "Delete a flashcard set by ID"
      tags:
        - "FlashCard Sets"
      responses:
        "204":
          description: "The flashcard set was deleted"
        "404":
          description: "The flashcard set was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /users/{userID}/collections/{collectionID}:
    parameters:
      - name: userID
        in: path
        required: true
        description: "The ID of the user"
        schema:
          type: "string"
      - name: collectionID
        in: path
        required: true
        description: "The ID of the collection"
        schema:
          type: "string"
    get:
      summary: "Get a flashcard set collection by ID"
      tags:
        - "Collections"
      responses:
        "200":
          description: "The flashcard set collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: "The flashcard set collection was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    put:
      summary: "Update a flashcard set collection by ID"
      tags:
        - "Collections"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                comment:
                  type: "string"
                  nullable: false
                  example: "I love this set!"
                sets:
                  type: "array"
                  nullable: true
                  items:
                    type: "number"
                    description: "The IDs of member flashcards"
                    example: 1
      
      responses:
        "200":
          description: "The updated flashcard set collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: "The flashcard set collection was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: "You are not authorized ('logged in') to update this collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: "You are not allowed to update this collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    delete:
      summary: "Delete a flashcard set collection by ID"
      tags:
        - "Collections"
      responses:
        "204":
          description: "The flashcard set collection was deleted"
        "404":
          description: "The flashcard set collection was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: "You are not authorized ('logged in') to delete this collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        '403':
          description: "You are not allowed to delete this collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /collections:
    get:
      summary: "Get all flashcard set collections"
      tags:
        - "Collections"
      responses:
        "200":
          description: "A list of flashcard set collections"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Collection"
    post:
      summary: "Create a new flashcard set collection"
      tags:
        - "Collections"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                comment:
                  type: "string"
                  nullable: false
                  example: "I love this set!"
                sets:
                  type: "array"
                  nullable: true
                  items:
                    type: "number"
                    description: "The IDs of member flashcards"
                    example: 1

      responses:
        "201":
          description: "The created flashcard set collection"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: "The flashcard set was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /collections/random:
    get:
      summary: "Redirect to a random flashcard set collection"
      tags:
        - "Collections"
      responses:
        "302":
          description: "Redirect to a random flashcard set collection"
          headers:
            Location:
              schema:
                type: "string"
                example: "/users/1/collections/1"
        "404":
          description: "There are no flashcard set collections"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
components:
  schemas:
...
    FlashCard:
      type: "object"
      properties:
        question:
          type: "string"
          example: "What is the capital of France?"
        answer:
          type: "string"
          example: "Paris"
        difficulty:
          $ref: "#/components/schemas/Difficulty"
          nullable: true
    FlashcardSet:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        name:
          type: "string"
          example: "European Capitals"
        cards:
          type: "array"
          items:
            $ref: "#/components/schemas/FlashCard"
        created_at:
          type: "string"
          format: "date-time"
          example: "2024-10-14T12:00:00Z"
        updated_at:
          type: "string"
          format: "date-time"
          example: "2024-10-15T12:00:00Z"
    ProblemDetails:
      type: "object"
      properties:
        title:
          type: "string"
          nullable: false
          description: "Summary of the problem"
        status:
          type: "number"
          nullable: false
          description: "HTTP status code"
        detail:
          type: "string"
          nullable: true
          description: "Additional problem details"
    Difficulty:
      type: "string"
      enum:
        - "easy"
        - "medium"
        - "hard"
    User:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        username:
          type: "string"
          example: "johndoe"
        admin:
          type: "boolean"
          example: false
          default: false
          nullable: true
          description: "The set creation limit should only apply to non-admin users"
...
    Collection:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        comment:
          type: "string"
          example: "I like these sets"
        owner:
          $ref: "#/components/schemas/User"
        sets:
          type: "array"
          items:
            $ref: "#/components/schemas/FlashcardSet"